findFilesWithText.sh - case insensitive option and some logging directoryStack.sh - some best practice fixes system.sh - method to clean a path for windows jira/lib/jira_lib.sh - silent curl output

+++ visual_studio/updateAssemblyAttribute.sh
@@ -10,11 +10,20 @@
 set -u #//error on unset variable
 set -e #//exit on error
 
+# echo print colors
+RED='\033[0;31m'
+
 # Path to project file of project being tested
 PROJECT_FILE="path/to/project.csproj"
 
 # name of the test project
 TEST_PROJECT="UnitTestProject"
+
+# Ensure $TEST_PROJECT is defined
+if [ -z "$TEST_PROJECT" ]; then
+  echo -e "${RED}Error: TEST_PROJECT variable is not defined.${NC}"
+  exit 1
+fi
 
 # Check if project file has already been updated
 if grep -q "$TEST_PROJECT" "$PROJECT_FILE"; then
@@ -24,6 +33,12 @@ fi
 
 # Create a temporary file for content
 TEMP_FILE=$(mktemp)
+
+# Check if temporary file was created successfully
+if [ ! -f "$TEMP_FILE" ]; then
+  echo -e "${RED}Error: Failed to create a temporary file.${NC}"
+  exit 1
+fi
 
 # Add the XML content to the temporary file (avoid issues with git bash newline)
 cat <<EOL > "$TEMP_FILE"
